+++ authority-portal-frontend/src/app/app.module.ts
@@ -6,6 +6,8 @@ import {BrowserModule} from '@angular/platform-browser';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {NgxsReduxDevtoolsPluginModule} from '@ngxs/devtools-plugin';
 import {NgxsModule} from '@ngxs/store';
+import {ErrorElementModule} from 'src/app/common/components/error-element/error-element.module';
+import {LoadingElementModule} from 'src/app/common/components/loading-element/loading-element.module';
 import {MaterialModule} from 'src/app/common/material/material.module';
 import {AuthorityInviteNewOrganizationModule} from 'src/app/pages/authority-invite-new-organization/authority-invite-new-organization.module';
 import {AuthorityOrganizationUserDetailPageModule} from 'src/app/pages/authority-organization-user-detail-page/authority-organization-user-detail-page.module';
@@ -60,6 +62,8 @@ import {SpConnectorListPageModule} from './pages/sp-connector-list-page/sp-conne
     NgxsReduxDevtoolsPluginModule.forRoot(),
     // Authority Portal
     PortalLayoutModule,
+    LoadingElementModule,
+    ErrorElementModule,
     DashboardPageModule,
     DevUtilsModule,
     LoadingPageModule,