+++ authority-portal-frontend/src/app/pages/participant-own-connector-list-page/state/participant-own-connector-list-page-state-impl.ts
@@ -1,9 +1,12 @@
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
-import {ignoreElements, map, tap} from 'rxjs/operators';
-import {Action, Selector, State, StateContext} from '@ngxs/store';
+import {ignoreElements, map, switchMap, take, tap} from 'rxjs/operators';
+import {Action, State, StateContext, Store} from '@ngxs/store';
 import {ConnectorOverviewEntryDto} from '@sovity.de/authority-portal-client';
 import {ApiService} from 'src/app/core/api/api.service';
+import {GlobalState} from 'src/app/core/global-state/global-state';
+import {GlobalStateImpl} from 'src/app/core/global-state/global-state-impl';
+import {GlobalStateUtils} from 'src/app/core/global-state/global-state-utils';
 import {Fetched} from 'src/app/core/utils/fetched';
 import {GetOwnOrganizationConnectors} from './participant-own-connector-list-page-actions';
 import {
@@ -17,13 +20,20 @@ import {
 })
 @Injectable()
 export class ParticipantOwnConnectorListPageStateImpl {
-  constructor(private apiService: ApiService) {}
+  constructor(
+    private apiService: ApiService,
+    private store: Store,
+    private globalStateUtils: GlobalStateUtils,
+  ) {}
 
   @Action(GetOwnOrganizationConnectors)
   onGetOwnOrganizationConnectors(
     ctx: StateContext<ParticipantOwnConnectorListPageState>,
   ): Observable<never> {
-    return this.apiService.getOwnOrganizationConnectors().pipe(
+    return this.globalStateUtils.getDeploymentEnvironmentId().pipe(
+      switchMap((deploymentEnvironmentId) =>
+        this.apiService.getOwnOrganizationConnectors(deploymentEnvironmentId),
+      ),
       map((result) => result.connectors),
       Fetched.wrap({failureMessage: 'Failed loading connectors'}),
       tap((connectors) => this.connectorsRefreshed(ctx, connectors)),