+++ authority-portal-frontend/src/app/pages/participant-register-own-connector-page/participant-register-own-connector-page.module.ts
@@ -5,6 +5,7 @@ import {BrowserModule} from '@angular/platform-browser';
 import {RouterModule} from '@angular/router';
 import {NgxsFormPluginModule} from '@ngxs/form-plugin';
 import {NgxsModule} from '@ngxs/store';
+import {DevUtilsModule} from 'src/app/common/components/dev-utils/dev-utils.module';
 import {PortalLayoutModule} from 'src/app/common/layouts/portal-layout/portal-layout.module';
 import {ParticipantRegisterOwnConnectorPageComponent} from './participant-register-own-connector-page/participant-register-own-connector-page.component';
 import {ParticipantRegisterOwnConnectorPageStateImpl} from './state/participant-register-own-connector-page-state-impl';
@@ -21,6 +22,8 @@ import {ParticipantRegisterOwnConnectorPageStateImpl} from './state/participant-
     ReactiveFormsModule,
     NgxsFormPluginModule,
     NgxsModule.forFeature([ParticipantRegisterOwnConnectorPageStateImpl]),
+
+    DevUtilsModule,
   ],
 })
 export class ParticipantRegisterOwnConnectorPageModule {}