+++ authority-portal-frontend/src/app/pages/provide-connector/provide-connector.module.ts
@@ -5,6 +5,7 @@ import {BrowserModule} from '@angular/platform-browser';
 import {RouterModule} from '@angular/router';
 import {NgxsFormPluginModule} from '@ngxs/form-plugin';
 import {NgxsModule} from '@ngxs/store';
+import {DevUtilsModule} from 'src/app/common/components/dev-utils/dev-utils.module';
 import {PortalLayoutModule} from 'src/app/common/layouts/portal-layout/portal-layout.module';
 import {ProvideConnectorPageComponent} from './provide-connector/provide-connector-page.component';
 import {ProvideConnectorPageStateImpl} from './state/provide-connector-page-state-impl';
@@ -17,6 +18,7 @@ import {ProvideConnectorPageStateImpl} from './state/provide-connector-page-stat
     RouterModule,
 
     PortalLayoutModule,
+    DevUtilsModule,
     FormsModule,
     ReactiveFormsModule,
     NgxsFormPluginModule,