+++ authority-portal-frontend/src/app/pages/sp-connector-list-page/state/sp-connector-list-page-state-impl.ts
@@ -1,9 +1,11 @@
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
-import {ignoreElements, map, tap} from 'rxjs/operators';
-import {Action, Selector, State, StateContext} from '@ngxs/store';
+import {ignoreElements, map, switchMap, take, tap} from 'rxjs/operators';
+import {Action, Selector, State, StateContext, Store} from '@ngxs/store';
 import {ConnectorOverviewEntryDto} from '@sovity.de/authority-portal-client';
 import {ApiService} from 'src/app/core/api/api.service';
+import {GlobalState} from 'src/app/core/global-state/global-state';
+import {GlobalStateImpl} from 'src/app/core/global-state/global-state-impl';
 import {Fetched} from 'src/app/core/utils/fetched';
 import {GetOwnOrganizationConnectors} from './sp-connector-list-page-actions';
 import {
@@ -17,13 +19,20 @@ import {
 })
 @Injectable()
 export class SpConnectorListPageStateImpl {
-  constructor(private apiService: ApiService) {}
+  constructor(private apiService: ApiService, private store: Store) {}
 
   @Action(GetOwnOrganizationConnectors)
   onGetOwnOrganizationConnectors(
     ctx: StateContext<SpConnectorListPageState>,
   ): Observable<never> {
-    return this.apiService.getOwnOrganizationConnectors().pipe(
+    return this.store.select<GlobalState>(GlobalStateImpl).pipe(
+      take(1),
+      map((globalState) => globalState.selectedEnvironment),
+      switchMap((selectedEnvironment) =>
+        this.apiService.getOwnOrganizationConnectors(
+          selectedEnvironment?.environmentId,
+        ),
+      ),
       map((result) => result.connectors),
       Fetched.wrap({failureMessage: 'Failed loading connectors'}),
       tap((connectors) => this.connectorsRefreshed(ctx, connectors)),