+++ authority-portal-frontend/src/app/shared/components/organization-detail/organization-detail.component.ts
@@ -1,5 +1,9 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {OrganizationDetailsDto} from '@sovity.de/authority-portal-client';
+import {
+  OrganizationDetailsDto,
+  OwnOrganizationDetailsDto,
+} from '@sovity.de/authority-portal-client';
+import {GlobalStateUtils} from 'src/app/core/global-state/global-state-utils';
 import {
   getOrganizationRegistrationStatusClasses,
   getOrganizationUserRegistrationStatusClasses,
@@ -14,10 +18,26 @@ export enum OrganizationActions {
   selector: 'app-organization-detail',
   templateUrl: './organization-detail.component.html',
 })
-export class OrganizationDetailComponent {
-  @Input() organization!: OrganizationDetailsDto;
+export class OrganizationDetailComponent implements OnInit {
+  @Input() organization!: OrganizationDetailsDto | OwnOrganizationDetailsDto;
   @Input() type!: 'Participant' | 'Authority';
   @Output() actions = new EventEmitter<OrganizationActions>();
+
+  dashboard!: {
+    dataOfferCount: number;
+    connectorCount: number;
+    memberCount: number;
+  };
+
+  ngOnInit() {
+    this.dashboard = {
+      dataOfferCount: (this.organization as OrganizationDetailsDto)
+        .dataOfferCount,
+      connectorCount: (this.organization as OrganizationDetailsDto)
+        .connectorCount,
+      memberCount: (this.organization as OrganizationDetailsDto).memberCount,
+    };
+  }
 
   // html doesn't see this function if it's just imported
   getOrganizationRegistrationStatusClasses =