+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Initial Authority Portal MVP Release with a minimalistic feature set.
 - Updated organization detail page to show user list and some stats regarding the organization
 - Added page for showing User's details
 - Added endpoint to retrieve available deployment environments
+- Added support for multiple environments
 
 
 #### Patch